+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix class methods not getting wrapped when using the decorator
+
 ## [1.4.3] - 2019-06-25
 
 ### Fixed