+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [1.4.4] - 2019-06-27
+
 ### Fixed
 
 - Fix class methods not getting wrapped when using the decorator